Moving results types to domain model
